refactor : 도메인 패키징 변경
