+++ src/main/java/nextstep/ladder/LadderGame.java
@@ -1,6 +1,10 @@
 package nextstep.ladder;
 
-import nextstep.ladder.domain.*;
+import nextstep.ladder.domain.ladderline.LadderGenerator;
+import nextstep.ladder.domain.ladderline.LadderLine;
+import nextstep.ladder.domain.point.RightPointRandomStrategy;
+import nextstep.ladder.domain.user.UserGenerator;
+import nextstep.ladder.domain.user.Users;
 import nextstep.ladder.view.InputView;
 import nextstep.ladder.view.ResultView;
