+++ src/main/java/nextstep/ladder/domain/point/RightPointRandomStrategy.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.point;
 
 import java.util.Random;
