+++ src/main/java/nextstep/ladder/domain/point/RightPointStrategy.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.point;
 
 public interface RightPointStrategy {
