+++ src/main/java/nextstep/ladder/domain/result/Result.java
@@ -1,0 +1,4 @@
+package nextstep.ladder.domain.result;
+
+public class Result {
+}
