+++ src/test/java/nextstep/ladder/domain/LadderLineTest.java
@@ -1,5 +1,7 @@
 package nextstep.ladder.domain;
 
+import nextstep.ladder.domain.ladderline.LadderGenerator;
+import nextstep.ladder.domain.ladderline.LadderLine;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
