+++ src/test/java/nextstep/ladder/domain/UserTest.java
@@ -1,5 +1,7 @@
 package nextstep.ladder.domain;
 
+import nextstep.ladder.domain.user.User;
+import nextstep.ladder.domain.user.UserGenerator;
 import nextstep.ladder.utils.LadderValidation;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
